backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]>=3.2.3
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.1
pydantic[email]>=2.9.2
requests>=2.32.3
groq>=0.11.0
openai>=1.54.4
